requirements.txt
pydantic
sqlAlchemy
uvicorn
pytz
fastapi
yagmail
keyring
verification-code
coverage
yagmail
keyring
PyYAML==5.4
requests==2.25.1
tqdm==4.56.0
Pygments==2.9.0
pyTelegramBotAPI
Flask
Flask-WTF
flask-cors
python-dateutil
